requirements.txt
pyspark
kafka-python
pandas
matplotlib
seaborn
jupyter
faker
dotenv
reportlab
streamlit
plotly
psycopg2-binary
python-dotenv
seaborn
numpy